ui: code gardening - renamed viewsBasic/integrations to integrationDefs and moved a few integration things from there to viewsBasic/tenantIntegrations
