+++ packages/app/src/client/viewsBasic/integrationDefs/cloudwatch/index.tsx
@@ -16,12 +16,12 @@
 
 import React from "react";
 
-import { IntegrationDef } from "client/viewsBasic/integrations/types";
+import { IntegrationDef } from "client/viewsBasic/integrationDefs/types";
 
-export const StackDriverIntegration: IntegrationDef = {
-  kind: "stack-driver",
+export const CloudwatchIntegration: IntegrationDef = {
+  kind: "cloudwatch",
   category: "exporter",
-  label: "Stack Driver",
+  label: "Cloudwatch",
   Form: () => <div />,
   Row: () => <div />,
   Card: () => <div />,
