+++ packages/app/src/client/viewsBasic/integrationDefs/k8sMetrics/commandTemplate.ts
