+++ packages/app/src/client/viewsBasic/integrationDefs/k8sMetrics/configTemplate.ts
